Initial Push
working but in a fragile state. currently you can only login to an
existing user and if succesful loads the map fragment. On error returns
a toast message with the error. The backend supports registering a user
we just need to implement the fragment for it and the logic for user
input.